test: add test_cost_model for roofline math + GPU preset sanity
6 groups validate the CostModel public API:
- Presets (B200, H100, A100, MI300X) match vendor specs to 1%.
- Ridge-point: peak_tflops × 1e3 / hbm_bw = FLOP/byte balance point.
- Wave efficiency: 0 → 0, exactly-filled → 1.0, half-filled → 0.5.
- SM occupancy: 32/64/128 regs/thread → 100/50/25%.
- validate_config rejects every failing constraint (C1-C8).
- Fusion benefit arithmetic: saved_ns, speedup, saved_launches.

Also fixes -Wdouble-promotion in CostModel.h's evaluate_cost loop
(float × hbm_bw was silently promoting to double).

+++ include/crucible/CostModel.h
@@ -469,15 +469,17 @@ struct CostBreakdown {
   float peak = hw.peak_tflops(dtype);
   if (peak > 0 && cb.wave_efficiency > 0 && cb.occupancy > 0) {
     double effective = static_cast<double>(peak) * 1e3
-                     * cb.wave_efficiency * cb.occupancy;
+                     * static_cast<double>(cb.wave_efficiency)
+                     * static_cast<double>(cb.occupancy);
     cb.compute_ns = static_cast<double>(flops) / effective;
   }
 
   // Memory time: latency + transfer
   // hbm_bw is in GB/s = bytes/ns (unit identity)
   if (hw.hbm_bw > 0) {
-    cb.memory_ns = hw.hbm_latency
-                 + static_cast<double>(bytes) / hw.hbm_bw;
+    cb.memory_ns = static_cast<double>(hw.hbm_latency)
+                 + static_cast<double>(bytes)
+                 / static_cast<double>(hw.hbm_bw);
   }
 
   // Total = max(compute, memory) + launch
